+++ CHANGELOG.md
@@ -4,10 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [unreleased]
+## [1.3.0] 2019-12-09
 
 ### Added
-- `filters` to spec
+- `filter` to specification
 - basic input filter types for `string`, `float`, `integer`, `boolean` and `ID`
 - `use function` for all PHP functions used
 - HTTP status codes to spec